refactor: update code view structure
